Refactor cv code
